chore: improve block directory structure
