+++ packages/umi-build-dev/src/plugins/commands/block/ui/client/locales/zh-CN.js
